backend/scraper/requirements.txt
# LangChain & LangGraph
langgraph==0.2.0
langchain==0.2.0
langchain-core==0.2.0

# Async & Task Queue
celery==5.3.4
redis==5.0.1

# Message Queue
pika==1.3.2

# Object Storage
minio==7.2.0

# Web Scraping with BeautifulSoup
beautifulsoup4==4.12.3
lxml==5.1.0
html5lib==1.1
requests==2.31.0
feedparser==6.0.11

# HTTP Client
httpx==0.26.0
urllib3==2.1.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Date/Time
python-dateutil==2.8.2

# Utilities
python-dotenv==1.0.0
PyYAML==6.0.1

# Monitoring
prometheus-client==0.19.0